requirements.txt
PyPDF2
aiohttp
python-dotenv
aiofiles
tensorflow
tensorflow_hub
